build: rename appConfig.ts to app-config.ts
Ref. metriport/metriport-internal#1634
