+++ packages/core/src/external/aws/app-config.ts
